refactor: separate sample data generator
